modularize test case generation script
