+++ test/samples/chart-config/scatterPlot.json
@@ -21,12 +21,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -50,12 +50,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -79,12 +79,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -108,12 +108,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -137,13 +137,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -167,13 +167,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -197,12 +197,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -226,12 +226,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -255,12 +255,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -284,12 +284,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -313,13 +313,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -343,13 +343,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -373,12 +373,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -402,12 +402,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -431,12 +431,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -460,12 +460,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -489,13 +489,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -519,13 +519,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -549,12 +549,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -578,12 +578,12 @@
       "marks": [
         {
           "type": "circle",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -607,12 +607,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -636,12 +636,12 @@
       "marks": [
         {
           "type": "line",
-
+		  
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -665,13 +665,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["Element"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -695,13 +695,13 @@
       "marks": [
         {
           "type": "text",
-
+		  
 		  "text": "$x",
 		  "per": ["group"]
         }
       ]
     },
-    "filename": "../data/elements.csv"
+    "dataPath": "../../data/elements_first_five.csv"
   },
 
 
@@ -729,7 +729,7 @@
 				"per": ["group"]
 			}]
 		},
-		"filename": "../data/elements.csv"
+		"dataPath": "../../data/elements_first_five.csv"
 	}
