--- test/samples/chart-config/scripts/README.txt
@@ -1,10 +1,0 @@
-templates: contains jinja template for generating visual test cases
-
-chart_gen: python file to run (will create an output testConfigs.js)
-
-		  
-Instructions: 
-1.  Install Python
-2.  Install jinja2 (Ensure Python recognizes the jinja module)
-   ^ Don't spend more than 15 min on the first two instructions like I did lol
-3.  Open the python IDLE and run chart_gen.py
