+++ test/samples/chart-config/scripts/templates/scatterPlot_template.json
@@ -33,7 +33,7 @@
         }
       ]
     },
-    "dataPath": "../../data/elements_first_five.csv"
+    "filename": "../data/elements.csv"
   },
 
 
@@ -66,7 +66,7 @@
 				"per": ["group"]
 			}]
 		},
-		"dataPath": "../../data/elements_first_five.csv"
+    "filename": "../data/elements.csv"
 	}
